+++ bugsnag-android-core/src/androidTest/java/com/bugsnag/android/ConfigurationTest.java
@@ -10,7 +10,6 @@ import static org.junit.Assert.assertTrue;
 import android.support.test.filters.SmallTest;
 import android.support.test.runner.AndroidJUnit4;
 
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -36,30 +35,6 @@ public class ConfigurationTest {
 
         assertEquals(notify, config.getEndpoints().getNotify());
         assertEquals(sessions, config.getEndpoints().getSessions());
-    }
-
-    @SuppressWarnings("deprecation")
-    @Test
-    public void testEndpoint() {
-        // Default endpoints
-        assertEquals("https://notify.bugsnag.com", config.getEndpoint());
-
-        // Setting an endpoint
-        String endpoint = "http://localhost:8000";
-        config.setEndpoint(endpoint);
-        assertEquals(endpoint, config.getEndpoint());
-    }
-
-    @SuppressWarnings("deprecation")
-    @Test
-    public void testSessionEndpoint() {
-        // Default endpoints
-        assertEquals("https://sessions.bugsnag.com", config.getSessionEndpoint());
-
-        // Setting an endpoint
-        String endpoint = "http://localhost:8000";
-        config.setSessionEndpoint(endpoint);
-        assertEquals(endpoint, config.getSessionEndpoint());
     }
 
     @Test
@@ -90,25 +65,6 @@ public class ConfigurationTest {
         // Should ignore when added to ignoreClasses
         config.setIgnoreClasses(new String[]{className});
         assertTrue(config.shouldIgnoreClass(className));
-    }
-
-    @SuppressWarnings("deprecation")
-    @Test
-    public void testInProject() {
-        // Shouldn't be inProject if projectPackages hasn't been set
-        assertFalse(config.inProject("com.bugsnag.android.Example"));
-
-        // Should be inProject if class in projectPackages
-        config.setProjectPackages(new String[]{"com.bugsnag.android"});
-        assertTrue(config.inProject("com.bugsnag.android.Example"));
-
-        // Shouldn't be inProject if class not in projectPackages
-        config.setProjectPackages(new String[]{"com.bugsnag.android"});
-        assertFalse(config.inProject("java.io.IOException"));
-
-        // Should be inProject if class is in projectPackages with null element
-        config.setProjectPackages(new String[]{null, "java.io.IOException"});
-        assertTrue(config.inProject("java.io.IOException"));
     }
 
     @Test
@@ -189,7 +145,7 @@ public class ConfigurationTest {
         Delivery delivery = BugsnagTestUtils.generateDelivery();
         configuration.setDelivery(delivery);
 
-        assertFalse(configuration.getDelivery() instanceof DeliveryCompat);
+        assertFalse(configuration.getDelivery() instanceof DefaultDelivery);
         assertEquals(delivery, configuration.getDelivery());
     }