+++ bugsnag-android-core/src/main/java/com/bugsnag/android/Configuration.java
@@ -3,7 +3,6 @@ package com.bugsnag.android;
 import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.text.TextUtils;
 
 import java.lang.reflect.Field;
 import java.util.Arrays;
@@ -48,7 +47,6 @@ public class Configuration extends Observable implements Observer {
 
     private boolean detectAnrs = false;
     private boolean detectNdkCrashes;
-    private long anrThresholdMs = 5000;
 
     @NonNull
     private MetaData metaData;
@@ -165,14 +163,6 @@ public class Configuration extends Observable implements Observer {
      * @deprecated use {@link Configuration#setEndpoints(Endpoints)}
      */
     @Deprecated
-    public void setEndpoint(@NonNull String endpoint) {
-        setEndpoints(new Endpoints(endpoint, getEndpoints().getSessions()));
-    }
-
-    /**
-     * @deprecated use {@link Configuration#setEndpoints(Endpoints)}
-     */
-    @Deprecated
     public void setEndpoints(@NonNull String notify, @NonNull String sessions) {
         setEndpoints(new Endpoints(notify, sessions));
     }
@@ -208,14 +198,6 @@ public class Configuration extends Observable implements Observer {
     @NonNull
     public String getSessionEndpoint() {
         return endpoints.getSessions();
-    }
-
-    /**
-     * @deprecated use {@link Configuration#setEndpoints(Endpoints)}
-     */
-    @Deprecated
-    public void setSessionEndpoint(@NonNull String endpoint) {
-        setEndpoints(new Endpoints(getEndpoints().getNotify(), endpoint));
     }
 
     /**
@@ -406,17 +388,6 @@ public class Configuration extends Observable implements Observer {
      */
     public void setEnableExceptionHandler(boolean enableExceptionHandler) {
         this.enableExceptionHandler = enableExceptionHandler;
-    }
-
-    /**
-     * Get whether or not User sessions are captured automatically.
-     *
-     * @return true if sessions are captured automatically
-     * @deprecated use {@link #getAutoCaptureSessions()}
-     */
-    @Deprecated
-    public boolean shouldAutoCaptureSessions() {
-        return getAutoCaptureSessions();
     }
 
     /**
@@ -684,38 +655,6 @@ public class Configuration extends Observable implements Observer {
     }
 
     /**
-     * @return the threshold at which ANRs are detected, in ms
-     * @see #setAnrThresholdMs(long)
-     *
-     * @deprecated This option has been superseded by the ANR threshold of the underlying OS
-     */
-    @Deprecated
-    public long getAnrThresholdMs() {
-        return anrThresholdMs;
-    }
-
-    /**
-     * Sets the time in milliseconds at which an
-     * <a href="https://developer.android.com/topic/performance/vitals/anr">ANR</a> is detected
-     * by Bugsnag. By default, Bugsnag will record an ANR whenever the main thread has been blocked
-     * for 5000 milliseconds or longer.
-     * <p/>
-     * If you wish to enable ANR detection, you should set the {@link #setDetectAnrs(boolean)}
-     * property to true.
-     * <p/>
-     * Attempting to set this property to any value below 1000ms will result in the anrThresholdMs
-     * being set as 1000ms.
-     *
-     * @param anrThresholdMs the threshold in ms at which ANRs should be detected
-     * @see #setDetectAnrs(boolean)
-     *
-     * @deprecated This option has been superseded by the ANR threshold of the underlying OS
-     */
-    @Deprecated
-    public void setAnrThresholdMs(long anrThresholdMs) {
-    }
-
-    /**
      * Supplies the headers which must be used in any request sent to the Error Reporting API.
      *
      * @return the HTTP headers
@@ -820,17 +759,6 @@ public class Configuration extends Observable implements Observer {
         if (!beforeRecordBreadcrumbTasks.contains(beforeRecordBreadcrumb)) {
             beforeRecordBreadcrumbTasks.add(beforeRecordBreadcrumb);
         }
-    }
-
-    /**
-     * Checks if the given class name should be marked as in the project or not
-     *
-     * @param className the class to check
-     * @return true if the class should be considered in the project else false
-     */
-    @Deprecated
-    protected boolean inProject(@NonNull String className) {
-        return Stacktrace.inProject(className, projectPackages);
     }
 
     /**