+++ bugsnag-android-core/src/main/java/com/bugsnag/android/NativeInterface.java
@@ -162,11 +162,6 @@ public class NativeInterface {
         NativeInterface.client = client;
     }
 
-    @Deprecated
-    public static void configureClientObservers(@NonNull Client client) {
-        setClient(client);
-    }
-
     @Nullable
     public static String getContext() {
         return getClient().getContext();
@@ -335,7 +330,8 @@ public class NativeInterface {
      */
     @SuppressWarnings("deprecation")
     public static void setSessionEndpoint(@NonNull final String endpoint) {
-        getClient().getConfig().setSessionEndpoint(endpoint);
+        Configuration config = getClient().getConfig();
+        config.setEndpoints(config.getEndpoint(), endpoint);
     }
 
     /**
@@ -343,7 +339,8 @@ public class NativeInterface {
      */
     @SuppressWarnings("deprecation")
     public static void setEndpoint(@NonNull final String endpoint) {
-        getClient().getConfig().setEndpoint(endpoint);
+        Configuration config = getClient().getConfig();
+        config.setEndpoints(endpoint, config.getSessionEndpoint());
     }
 
     /**